requirements.txt
asgiref==3.5.0
dj-database-url==0.5.0
Django==3.2.12
django-cors-headers==3.11.0
django-environ==0.8.1
django-filter==21.1
django-resized==0.3.11
djangorestframework==3.13.1
environ==1.0
gunicorn==20.1.0
Pillow==9.0.1
pkg_resources==0.0.0
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
